+++ packages/client/src/common/commonTypes.ts
@@ -4,8 +4,9 @@
  * ------------------------------------------------------------------------------------------ */
 
 import type { BaseLanguageClient } from 'vscode-languageclient/browser.js';
+import type { Socket } from 'socket.io-client';
 
-export type ConnectionConfigOptions = WebSocketConfigOptionsDirect | WebSocketConfigOptionsParams | WebSocketConfigOptionsUrl | WorkerConfigOptionsParams | WorkerConfigOptionsDirect;
+export type ConnectionConfigOptions = WebSocketConfigOptionsDirect | WebSocketConfigOptionsParams | WebSocketConfigOptionsUrl | WorkerConfigOptionsParams | WorkerConfigOptionsDirect | SocketIoConfigOptionsDirect;
 
 export interface WebSocketCallOptions {
     /** Adds handle on languageClient */
@@ -14,9 +15,12 @@ export interface WebSocketCallOptions {
     reportStatus?: boolean;
 }
 
-export interface WebSocketConfigOptionsDirect {
+export interface WebSocketLikeConfig {
+    webSocket: WebSocket | Socket;
+}
+
+export interface WebSocketConfigOptionsDirect extends WebSocketLikeConfig {
     $type: 'WebSocketDirect'
-    webSocket: WebSocket
     startOptions?: WebSocketCallOptions;
     stopOptions?: WebSocketCallOptions;
 }
@@ -57,4 +61,8 @@ export interface WorkerConfigOptionsDirect {
     $type: 'WorkerDirect';
     worker: Worker;
     messagePort?: MessagePort;
+}
+
+export interface SocketIoConfigOptionsDirect extends WebSocketLikeConfig {
+    $type: 'SocketIoDirect'
 }