+++ vitest.config.ts
@@ -53,14 +53,15 @@ export const vitestConfig = {
             // '**/client/test/editorApp/editorApp.noservices.test.ts',
             // '**/client/test/editorApp/editorApp.wrongservices.test.ts',
             // '**/client/test/editorApp/config.test.ts',
+            '**/socketio/test/browser/client.test.ts',
             // '**/wrapper-react/test/index.test.tsx',
             // '**/wrapper-react/test/index.extapi.test.tsx',
             // '**/wrapper-react/test/index.strictmode.test.tsx',
             // '**/wrapper-react/test/index.lc.test.tsx',
             // '**/wrapper-react/test/index.lc.strictmode.test.tsx',
             // '**/wrapper-react/test/index.viewsservice.test.tsx',
-            '**/client/test/**/*',
-            '**/wrapper-react/test/**/*'
+            // '**/client/test/**/*',
+            // '**/wrapper-react/test/**/*'
         ],
         exclude: [
             '**/support/**/*',